DBZ-1478 Renamed deployment package
