+++ debezium-quarkus/extensions/outbox/deployment/src/main/java/io/debezium/outbox/quarkus/DebeziumOutboxConfig.java
@@ -3,7 +3,7 @@
  *
  * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
  */
-package io.debezium.quarkus.outbox;
+package io.debezium.outbox.quarkus;
 
 import io.quarkus.runtime.annotations.ConfigItem;
 import io.quarkus.runtime.annotations.ConfigPhase;
